+++ Plugin.cs
@@ -1,8 +1,18 @@
 using BepInEx;
 using BepInEx.Logging;
+using LethalLib.Extras;
+using LethalLib.Modules;
+using System.Collections.Generic;
+using System;
 using System.IO;
 using System.Reflection;
+using System.Text;
 using UnityEngine;
+using System.Collections;
+using JuicesMod.Behaviours;
+using GameNetcodeStuff;
+using JuicesMod.Properties;
+using System.Linq;
 
 namespace JuicesMod
 {
@@ -15,7 +25,7 @@ namespace JuicesMod
         internal new static ManualLogSource Logger;
         internal static Config BoundConfig { get; private set; } = null;
 
-        private string[] FRUITS = ["Orange", "Apple", "Pineapple", "Tomato", "Prune"];
+        public JuicesBuilder JuicesBuilder { get; private set; } = null;
 
         private void Awake()
         {
@@ -26,20 +36,53 @@ namespace JuicesMod
             string assetDir = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "juicesmod");
             AssetBundle bundle = AssetBundle.LoadFromFile(assetDir);
 
-            JuicesBuilder juicesBuilder = new JuicesBuilder(bundle);
+            RegisterRPCs();
+            JuiceTypeProperty.Initialize();
 
-            foreach(string fruit in FRUITS)
+            JuicesBuilder = new JuicesBuilder(bundle);
+
+            foreach (JuiceTypeProperty type in JuiceTypeProperty.JUICE_TYPES.Values)
             {
-                juicesBuilder.registerCarton($"{fruit}Juice001Item");
-                juicesBuilder.registerPremium($"{fruit}Juice002Item");
+                foreach (FruitProperty fruit in FruitProperty.FRUITS)
+                {
+                    JuicesBuilder.registerJuice(fruit, type);
+                }
+                JuicesBuilder.registerMultifruitJuice(type);
             }
 
-            //TerminalNode node = ScriptableObject.CreateInstance<TerminalNode>();
-            //node.clearPreviousText = true;
-            //node.displayText = "\"Le jus d'orange ça pique !\"\nDaiick - 2020\n\n";
-            //Items.RegisterShopItem(orangeJuice001, null, null, node, 0);
+            UnlockableItemDef juiceBlender = bundle.LoadAsset<UnlockableItemDef>("Assets/JuicesMod/ShipJuiceBlenderItem.asset");
+            if (juiceBlender?.unlockable?.prefabObject != null)
+            {
+                NetworkPrefabs.RegisterNetworkPrefab(juiceBlender.unlockable.prefabObject);
+                Utilities.FixMixerGroups(juiceBlender.unlockable.prefabObject);
+            }
+            if (juiceBlender != null)
+            {
+                Unlockables.RegisterUnlockable(juiceBlender, 80, StoreType.ShipUpgrade);
+            }
+            else
+            {
+                Logger.LogError("Unable to load JUICE BLENDER !!!");
+            }
 
-            Logger.LogInfo("Juices Mod");
+            Logger.LogInfo("Juices Mod Loaded !");
+        }
+
+        private void RegisterRPCs()
+        {
+            Type[] types = Assembly.GetExecutingAssembly().GetTypes();
+            foreach (Type type in types)
+            {
+                MethodInfo[] methods = type.GetMethods(BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static);
+                foreach (MethodInfo method in methods)
+                {
+                    object[] attributes = method.GetCustomAttributes(typeof(RuntimeInitializeOnLoadMethodAttribute), false);
+                    if (attributes.Length > 0)
+                    {
+                        method.Invoke(null, null);
+                    }
+                }
+            }
         }
     }
 }